+++ GUI/MainMenu.h
@@ -13,7 +13,7 @@
 #include <fstream>
 #include <iostream>
 #include "../ProjectSourceCode/Map.h"
-#include "../ProjectSourceCode/MapBuilder.h"
+#include "../ProjectSourceCode/Builder/MapBuilder.h"
 
 
 namespace fs = std::filesystem;
