+++ [COMP345] Assignment 1.vcxproj.filters
@@ -87,7 +87,7 @@
     <ClCompile Include="ProjectSourceCode\Interactable\Interactable.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ProjectSourceCode\MapBuilder.cpp">
+    <ClCompile Include="ProjectSourceCode\Builder\MapBuilder.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="ProjectSourceCode\Interactable\Wall.cpp">
@@ -212,7 +212,7 @@
     <ClInclude Include="ProjectSourceCode\Interactable\Interactable.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ProjectSourceCode\MapBuilder.h">
+    <ClInclude Include="ProjectSourceCode\Builder\MapBuilder.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ProjectSourceCode\Interactable\Wall.h">
